type annotation of the return value is fixed

+++ 0x00-python_variable_annotations/2-floor.py
@@ -2,14 +2,14 @@
 """Basic annotations -floor"""
 
 
-def floor(n: float) -> float:
+def floor(n: float) -> int:
     """ type annotated function that takes n as argument
         and returns the floor of the float
 
         Args:
             n: float - the float number
         Returns:
-            floor: float - the floor of the number
+            floor: int - the floor of the number
     """
 
-    return float(int(n))
+    return int(n)
